refactor(wardens): extract CI-status capsule from gate engine (LIA-306)
Step 3 of decomposing the ~4000-line codex_warden_hooks.py. Moves the
GitHub Actions CI-status cluster — the _CI_STATUS_*/_BUCKET_* constants and
_query_gh_checks, _check_ci_status, _ci_block_reason — verbatim into a new
warden_hooks/ci_status.py.

Pure leaf (stdlib only: subprocess + json), so plain re-export with NO
injection seam — unlike the verdict-store capsule. _check_ci_status is
monkeypatched by tests, but its callers (approve_admin_merge,
run_admin_merge_gate) stay in the entry referencing the re-exported name, so
the monkeypatch still rebinds what they see. All 9 constants + 3 functions
re-exported so hooks.<name> resolves for callers + tests.

Also corrects the warden_hooks package docstring, which claimed every capsule
is a pure zero-coupling leaf — inaccurate since verdict_store introduced the
injection seam; now describes both kinds.

Entry: -146 net (4000 -> 3854). No behavior change.
